feat(evm): adjust decimals to match evm (#666)
* support uint256 serialization

* update block and tx serde

* wip

* format currency

* rename to WEI

* update api sync

* milestones

* genesis block

* fix test configs

* add dependency

* update custom proposal

* revert comment

* style: resolve style guide violations

* update test tx builders

---------

+++ packages/api-database/source/models/block.ts
@@ -54,19 +54,19 @@ export class Block {
 
 	@Column({
 		nullable: false,
-		type: "bigint",
+		type: "numeric",
 	})
 	public readonly totalAmount!: string;
 
 	@Column({
 		nullable: false,
-		type: "bigint",
+		type: "numeric",
 	})
 	public readonly totalFee!: string;
 
 	@Column({
 		nullable: false,
-		type: "bigint",
+		type: "numeric",
 	})
 	public readonly reward!: string;
